% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BigData_toy.R
\docType{data}
\name{big_data_toy}
\alias{big_data_toy}
\title{A Toy Claims \code{tibble} for Performance Demonstration}
\format{
A \code{tibble} with 100,000 rows and 6 variables:
\describe{
  \item{ID}{patient ID to be grouped by}
  \item{date}{date of claim}
  \item{days_supply}{number of days supplied}
}
}
\usage{
big_data_toy
}
\description{
A toy pharmaceutical claims data set meant to be used to benchmark other algorithms
}
\keyword{datasets}
