% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calculations.R
\name{calculate_pdc}
\alias{calculate_pdc}
\title{Calculate Proportion Days Covered}
\usage{
calculate_pdc(.data, .max_date, .min_date, .summarise = TRUE)
}
\arguments{
\item{.data}{data frame}

\item{.max_date}{Character in year-month-date format, maximum date in consideration for adherence}

\item{.min_date}{Character in year-month-date format, minimum date in consideration for adherence}

\item{.summarise}{Logical value (defaulting to TRUE) indicating whether the output should be summarised or not}
}
\value{
a summarised tibble, by default, with percentage of days a patient had medication coverage
}
\description{
Calculate the proportion of days covered from a pharmaceutical claims database. This function is suggested only  
after one has properly adjusted their dates (\code{propagate_dates()}) and identified gaps in therapy
(\code{identify_gaps()}). Note that a proportion of days covered will never be greater than 1 and has a philosophy that 
overlapping fill dates should never count for gaps prior to.
}
\examples{
library(adheRenceRX)
library(dplyr)

toy_claims \%>\% 
  group_by(ID) \%>\% 
  propagate_date() \%>\% 
  identify_gaps() \%>\% 
  calculate_pdc(.max_date = "2020-07-28", .min_date = "2020-01-01")

}
