% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dates.R
\name{date_check}
\alias{date_check}
\title{Restructuring Dates to Remove Overlap}
\usage{
date_check(df)
}
\arguments{
\item{df}{a claims data frame with a date of a medication claim and the corresponding days supply}
}
\value{
A new claims data frame with an appended column, \code{adjusted_date}
}
\description{
This is a function meant to be utilized within propagate_date() in order to adjust pharmaceutical claims dates to prevent
overlapping in adherence calculations, per Canfield SL, Zuckerman A, Anguiano RH, Jolly JA, DeClercq J.
Navigating the wild west of medication adherence reporting in specialty pharmacy. J Manag Care Spec Pharm. 2019;25(10):1073-77.
}
