% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gaps.R
\name{identify_gaps}
\alias{identify_gaps}
\title{Identify Gaps in A Patient's Prescription Claims History}
\usage{
identify_gaps(.data)
}
\arguments{
\item{.data}{data frame}
}
\value{
A new claims tibble with an appended column, "gap"
}
\description{
Compute gaps in a patient's prescription claims history from the end of their prior fill 
after one has adjusted dates with \code{propagate_date()}. This function assumes that one has
arranged the dates and grouped appropriately outside of the function.
}
