% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propagate_date.R
\name{propagate_date}
\alias{propagate_date}
\title{Tidy Verb to Forward-Adjusted Overlapping Pharmaceutical Claim Dates}
\usage{
propagate_date(.data, date_var = date, days_supply_var = days_supply)
}
\arguments{
\item{.data}{Data to be piped into the function}

\item{date_var}{Date column (will default to 'date' if not specified)}

\item{days_supply_var}{Days supply column (will default to "days_supply" if none supplied)}
}
\value{
A new claims data frame with an appended column, "adjusted_date"
}
\description{
To push supposedly overlapping dates forward for a specified group (e.g. patient ids or medication classes).
Used directly after a claims data frame has been grouped appropriately for adherence measurements per
Canfield SL, Zuckerman A, Anguiano RH, Jolly JA, DeClercq J.
Navigating the wild west of medication adherence reporting in specialty pharmacy. J Manag Care Spec Pharm. 2019;25(10):1073-77.
}
