% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Episodes.R
\name{rank_episodes}
\alias{rank_episodes}
\title{Rank Episodes of Care from Claims Data}
\usage{
rank_episodes(.data, .permissible_gap = NULL, .initial_rank = 1)
}
\arguments{
\item{.data}{Data frame with "gap", "initial_rank", and "permi_gap" columns appended from \code{identify_gaps()}}

\item{.permissible_gap}{Integer value suggesting the maximum gap allowed before labeling a new episode of care}

\item{.initial_rank}{Integer value to identify what the first rank should be (defaults to 1).}
}
\value{
a data frame with an "episode" column appended, which ranks episodes of care in time
}
\description{
This function identifies and labels all episodes of care for a given patient in chronological order. It is meant to be used after one has
appropriately adjusted dates (\code{propagate_date()}) and identified gaps (\code{identify_gaps()}) and at the minimum assumes that a data frame has 
an adjusted_date column, which identifies the date after being adjusted for overlaps, and a gap column, which identifies gaps in therapy. Both of these will be generated
by the aforementioned functions within this package!
}
\examples{
library(adheRenceRX)
library(dplyr)

toy_claims \%>\% 
  filter(ID == "D") \%>\% 
  propagate_date() \%>\% 
  identify_gaps() \%>\% 
  rank_episodes(.permissible_gap = 20, .initial_rank = 1)
  
}
