% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_gaps.R
\name{summarise_gaps}
\alias{summarise_gaps}
\title{Summarise Gaps in Therapy}
\usage{
summarise_gaps(.data)
}
\arguments{
\item{.data}{Data to be piped into the function}
}
\value{
A summary of gaps in therapy
}
\description{
This function serves as a convenience wrapper of dplyr::summarise(), which takes the grouped variables and
summarises their gaps in therapy. This function is to be used after propagate_date().
}
