% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Episodes.R
\name{rank_episodes}
\alias{rank_episodes}
\title{Rank Episodes of Care from Claims Data}
\usage{
rank_episodes(.data, .permissible_gap = NULL, .initial_rank = 1)
}
\arguments{
\item{.data}{Data frame with a "gap" column appended from \code{identify_gaps()}}

\item{.permissible_gap}{Integer value suggesting the maximum gap allowed before labeling a new episode of care}

\item{.initial_rank}{Integer value to identify what the indexing rank should be (defaults to 1).}
}
\value{
a data frame with an "episode" column appended, which ranks episodes of care in time
}
\description{
This function identifies and labels all episodes of care for a given patient in chronological order. A new episode begins after a specified gap in therapy has occurred. It is meant to be used after one has
appropriately adjusted dates (\code{propagate_date()}) and identified gaps (\code{identify_gaps()}).
}
\examples{
library(adheRenceRX)
library(dplyr)

toy_claims \%>\% 
  filter(ID == "D") \%>\% 
  propagate_date() \%>\% 
  identify_gaps() \%>\% 
  rank_episodes(.permissible_gap = 20, .initial_rank = 1)
  
}
